embulk-standards: Refactor TestCsvGuessPlugin
