--- embulk-standards/src/test/resources/org/embulk/standards/guess/csv/test/common.yml
@@ -1,1 +1,0 @@
-type: file
